refactor: Remove `react-native-paper`
`react-native-paper` has stopped keeping up to date with the latest React Native releases. In fairness to them, keeping up is a tall order, especially for the size of their component library and the number of places they hook into native for just that small extra bit of polish.

But, I need to keep my apps up to date, and I can't be held back by any third-party dependency. So, it's time for me to start rolling my own component library.

This refactor dramatically drops the visual appeal of the app -- I still need to do a lot of theming work. But, all the functionality is there!

Additionally, the new components are based only on core React Native, so they may even work on Desktop platforms (not a target platform for this app, but certainly for other apps I maintain).