feat: Add button to ignore tasks
"Delete" doesn't last if the task keeps getting completed. But we can ignore the task to acknowledge that the sync can see it, but denote that we don't want the sync to do anything.

Ignored tasks also get sorted to the bottom so they stay "out of sight, out of mind".